Moved to platform specific mix-ins
