+++ streams/js/src/main/scala/zio/stream/platform.scala
@@ -67,3 +67,5 @@ trait ZStreamPlatformSpecificConstructors { self: ZStream.type =>
 
 }
 trait StreamPlatformSpecificConstructors
+
+trait ZTransducerPlatformSpecificConstructors
