+++ streams/shared/src/main/scala/zio/stream/ZTransducer.scala
@@ -102,7 +102,7 @@ abstract class ZTransducer[-R, +E, -I, +O](val push: ZManaged[R, Nothing, Option
     ZTransducer[R1, E1, I, P](self.push.map(push => i => push(i).flatMap(_.mapM(f))))
 }
 
-object ZTransducer {
+object ZTransducer extends ZTransducerPlatformSpecificConstructors {
   def apply[R, E, I, O](
     push: ZManaged[R, Nothing, Option[Chunk[I]] => ZIO[R, E, Chunk[O]]]
   ): ZTransducer[R, E, I, O] =
